refactor(l1): move engine_only_context into ef_tests-engine crate
Relocates the shared-SyncManager static and per-thread merkle pool
helper out of `ethrex-rpc::test_utils` and into the engine crate that
owns them (`tooling/ef_tests/engine/src/engine_ctx.rs`). They only
exist to amortise per-fixture cost across the ~5600-fixture suite;
production has no reason to share a `SyncManager` across contexts.

Drops the `rayon` dep from `ethrex-rpc`.
