+++ tooling/ef_tests/engine/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod engine_ctx;
 pub mod exception_mapper;
 pub mod fixture;
 pub mod harness;
